requirements.txt
streamlit>=1.32.0
supabase>=2.10.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
plotly>=5.20.0
pytz>=2024.1
lxml>=5.1.0
httpx>=0.26.0
websockets>=13.0
